menyelesaikan tampilan home milik role student & refactoring beberapa component

+++ app/Http/Livewire/StudentAction.php
@@ -10,7 +10,7 @@ class StudentAction extends Component
 {
     use WithFileUploads;
 
-    public $name, $photo, $gender, $email, $class, $nis, $nisn, $address, $number_phone, $status, $id_user;
+    public $name, $photo, $gender, $email, $class, $nis, $nisn, $number_phone, $status, $id_user, $photo_profile;
 
     protected $listeners = [
         "visibleView"
@@ -35,17 +35,33 @@ class StudentAction extends Component
             $this->gender = $user->gender;
             $this->email = $user->email;
             $this->class = $user->profile->class->class;
-            $this->nis = $user->profile->NIS;
-            $this->nisn = $user->profile->NISN;
+            $this->nis = $user->profile->nis;
+            $this->nisn = $user->profile->nisn;
             $this->address = $user->profile->address;
             $this->number_phone = $user->profile->number_phone;
+            $this->photo_profile = $user->profile->photo_profile;
         }
     }
 
     public function destroy()
     {
         $user = User::find($this->id_user);
-        $session_data = $user->delete() ? ["success", "Berhasil Menghapus Data Siswa!"] : ["failed", "Terjadi Error saat menghapus Data Siswa!"];
+        try
+        {
+            $user->delete();
+
+            if(Storage::exists("public/photo_profile_student/{$user->profile->photo_profile}") && $user->profile->photo_profile !== "default.png")
+            {
+                Storage::delete("public/photo_profile_student/{$user->profile->photo_profile}");
+            }
+
+            $session_data = ["success", "Berhasil Menghapus Data Siswa!"];
+        }
+        catch(\Exception $ex)
+        {
+            $session_data = ["failed", "Terjadi Error saat menghapus Data Siswa!"];
+        }
+
 
         $this->emit("defaultView", $this->class, $session_data[0], $session_data[1]);
         $this->resetProp();
@@ -63,34 +79,41 @@ class StudentAction extends Component
 
     public function submitForm()
     {
-        $user = User::find($this->id_user);
-        if($this->getErrorBag()->count())
-        {
-            dd($this->getErrorBag());
-        }
-
+        $user = User::findOrFail($this->id_user);
         $this->validate(Profile::rules_student($user), Profile::messages_student());
 
+//        jika user meng-upload foto profil baru..
         if($this->photo)
         {
-            if(Storage::exists("public/photo_profile_student/{$user->profile->photo_profile}"))
+//            jika file photo_profile user yang lama ada & valuenya tidak default.png, maka hapus foto tsb
+            if(Storage::exists("public/photo_profile_student/{$user->profile->photo_profile}") && $user->profile->photo_profile !== "default.png")
             {
                 Storage::delete("public/photo_profile_student/{$user->profile->photo_profile}");
             }
 
+//            beri nama foto & store ke server
             $photo_name = uniqid() . $this->photo->getClientOriginalName();
             $this->photo->storeAs("photo_profile_student", $photo_name, "public");
 
+//            jika ada dir livewire-tmp, maka hapus
             if(Storage::exists("livewire-tmp"))
             {
                 Storage::deleteDirectory("livewire-tmp");
             }
         }
 
+/*        panggil static method updateProfile dari obj Profile untuk mengupdate profile, kirimkan 3 parameter,
+            parameter 1 berisi value var yang ada di objek ini(this),
+            parameter 2 berisi obj user yang akan diupdate,
+            parameter 3 berisi nilai photo_profile
+*/
         Profile::updateProfile(get_object_vars($this), $user, $photo_name ?? $user->profile->photo_profile);
         $this->status = "index";
         $this->photo = null;
         session()->flash("success", "Berhasil Mensunting Data Siswa!");
+
+//        setelah berhasil, kirimkan emit ke diri sendiri bernama visibleView dan mengirimkan var user sebagai argument
+        $this->emitSelf("visibleView", $user);
     }
 
     public function resetProp()
